Refactor AuthService and HttpDispatcherHandler for long-based user ID handling
- Updated AuthService to use long data type for user IDs, enhancing compatibility with the new Snowflake ID generation method.
- Refactored methods related to user registration, retrieval, and status changes to accommodate long user IDs.
- Modified HttpDispatcherHandler to handle long user IDs in HTTP requests, ensuring consistent data type usage across the application.
- Improved pending user management by implementing a more efficient key storage strategy in NeuroDB.

+++ chat-server/src/main/java/com/chat/network/http/HttpDispatcherHandler.java
@@ -362,7 +362,7 @@ public class HttpDispatcherHandler extends SimpleChannelInboundHandler<FullHttpR
             List<Long> pendingIds = authService.getPendingUserIds();
             List<Map<String, Object>> users = new ArrayList<>();
             for (Long id : pendingIds) {
-                User u = authService.getUserByUserId(id.intValue());
+                User u = authService.getUserByUserId(id);
                 if (u != null) {
                     users.add(Map.<String, Object>of(
                             "userId", u.getUserId(),
@@ -397,7 +397,7 @@ public class HttpDispatcherHandler extends SimpleChannelInboundHandler<FullHttpR
             sendJson(ctx, HttpResponseStatus.BAD_REQUEST, Map.of("error", "userId required"));
             return;
         }
-        int userId = ((Number) map.get("userId")).intValue();
+        long userId = ((Number) map.get("userId")).longValue();
         try {
             authService.approve(userId);
             sendJson(ctx, HttpResponseStatus.OK, Map.of("ok", true));
@@ -427,7 +427,7 @@ public class HttpDispatcherHandler extends SimpleChannelInboundHandler<FullHttpR
             sendJson(ctx, HttpResponseStatus.BAD_REQUEST, Map.of("error", "userId required"));
             return;
         }
-        int userId = ((Number) map.get("userId")).intValue();
+        long userId = ((Number) map.get("userId")).longValue();
         try {
             authService.reject(userId);
             sendJson(ctx, HttpResponseStatus.OK, Map.of("ok", true));
@@ -457,7 +457,7 @@ public class HttpDispatcherHandler extends SimpleChannelInboundHandler<FullHttpR
             sendJson(ctx, HttpResponseStatus.BAD_REQUEST, Map.of("error", "userId and action required"));
             return;
         }
-        int targetUserId = ((Number) map.get("userId")).intValue();
+        long targetUserId = ((Number) map.get("userId")).longValue();
         String action = (String) map.get("action");
         Number durationNum = (Number) map.get("durationHours");
         long durationMs = durationNum != null ? (long) (durationNum.doubleValue() * 3600 * 1000) : 0;
